Turbo Sport 1-70 outers

+++ script/thumbnails.py
@@ -17,21 +17,27 @@ def build(root_dir: str):
     else:
         for t_root, t_dirs, t_files in os.walk(output_dir):
             for filename in t_files:
-                if filename.endswith(".png"):
+                if filename != "missed.png" and filename.endswith(".png"):
                     os.remove(os.path.join(t_root, filename))
 
     for original in ["inner", "outer"]:
         os.makedirs(output_dir / original, exist_ok=True)
+        out_root = output_dir / original
         scan_path = SCANS_DIR / root_dir[len(str(ROOT)) + 1:] / original
-        for filename in os.listdir(scan_path):
-            if filename.endswith(".png"):
-                full_path = scan_path / filename
-                img = Image.open(full_path)
-                img.thumbnail(SIZE)
-                img.save(os.path.join(output_dir, original, filename))
+        for t_root, t_dirs, t_files in os.walk(scan_path):
+            x_path = Path(t_root).relative_to(scan_path)
+            for dirname in t_dirs:
+                os.makedirs(out_root / x_path / dirname, exist_ok=True)
+            file_root = out_root / x_path
+            for filename in t_files:
+                if filename.endswith(".png"):
+                    full_path = Path(t_root) / filename
+                    img = Image.open(full_path)
+                    img.thumbnail(SIZE)
+                    img.save(os.path.join(file_root, filename))
 
 
 if __name__ == "__main__":
-    for root, dirs, files in os.walk(f"../gum_wrappers/kent/turbo/sport/401-470"):
+    for root, dirs, files in os.walk(f"../gum_wrappers/kent/turbo/sport/1-70"):
         if "thumbnails" in dirs:
             build(root)
